+++ R/run_setup.R
@@ -44,7 +44,9 @@ run_setup <-
     }
 
     if (missing(conn)) {
+
       conn <- eval(rlang::parse_expr(conn_fun))
+
       on.exit(pg13::dc(conn = conn, verbose = verbose), add = TRUE,
               after = TRUE)
     }